Created FileOpenDialog to select a file from the file explorer

+++ GraphX-Rendering-Engine/src/entities/Camera.cpp
@@ -95,8 +95,8 @@ namespace engine
 				m_RenderStateDirty = true;
 
 				float distance = (CameraPosition - gm::Vector3(LastPosition, (CameraPosition.z == 0.0f ? -1.0f : 0.0f))).Magnitude();
-				float Yaw = gm::MathUtil::TanInverse(xOffset / distance);
-				float Pitch = gm::MathUtil::TanInverse(yOffset / distance);
+				float Yaw = (float) gm::MathUtil::TanInverse(xOffset / distance);
+				float Pitch = (float) gm::MathUtil::TanInverse(yOffset / distance);
 
 				gm::Vector3 Angles(-Pitch, -Yaw, 0);
 				EulerAngles += Angles;
